refactor(video): keep screenshot preparation hotspots within budget
