+++ src/content/video/videoScreenshotPreparationRequestStore.ts
@@ -3,33 +3,20 @@ import {
   createVideoScreenshotPreparationQueueStateSnapshot,
   type VideoScreenshotPreparationQueueStateSnapshot
 } from './videoScreenshotPreparationRequestSnapshots';
-
-export interface VideoScreenshotPreparationAttemptLike {
-  sourceVideo: HTMLVideoElement;
-  scope: {
-    dispose(): void;
-  };
-}
-
-interface VideoScreenshotPreparationRequestState<
-  TAttempt extends VideoScreenshotPreparationAttemptLike
-> {
-  tracked: boolean;
-  hiddenAttemptCount: number;
-  hiddenRetryAvailableAt: number;
-  hiddenTerminalFailed: boolean;
-  explicitVisible: boolean;
-  visibleAttempted: boolean;
-  visibleInFlight: boolean;
-  hiddenAttempt: TAttempt | null;
-}
-
-export interface VideoScreenshotHiddenAttemptFailureOptions {
-  maxAttempts: number;
-  retryAvailableAt: number;
-}
-
-export type VideoScreenshotHiddenAttemptFailureResult = 'ignored' | 'retry' | 'terminal';
+import {
+  canStartVideoScreenshotHiddenAttempt,
+  createVideoScreenshotPreparationRequestState,
+  disposeVideoScreenshotHiddenAttempt,
+  getNextVideoScreenshotHiddenRetryAt,
+  recordVideoScreenshotHiddenAttemptFailure,
+  resetVideoScreenshotHiddenProgress,
+  resetVideoScreenshotPreparationRequestState,
+  shouldRetainVideoScreenshotPreparationRequest,
+  type VideoScreenshotHiddenAttemptFailureOptions,
+  type VideoScreenshotHiddenAttemptFailureResult,
+  type VideoScreenshotPreparationAttemptLike,
+  type VideoScreenshotPreparationRequestState
+} from './videoScreenshotPreparationRequestState';
 
 export class VideoScreenshotPreparationRequestStore<
   TAttempt extends VideoScreenshotPreparationAttemptLike
@@ -45,7 +32,7 @@ export class VideoScreenshotPreparationRequestStore<
   track(captureId: string): void {
     const request = this.ensureRequest(captureId);
     request.tracked = true;
-    this.resetHiddenProgress(request);
+    resetVideoScreenshotHiddenProgress(request);
     this.emit();
   }
 
@@ -55,7 +42,7 @@ export class VideoScreenshotPreparationRequestStore<
       const request = this.ensureRequest(captureId);
       if (!request.tracked) {
         request.tracked = true;
-        this.resetHiddenProgress(request);
+        resetVideoScreenshotHiddenProgress(request);
         didChange = true;
       }
     }
@@ -69,8 +56,8 @@ export class VideoScreenshotPreparationRequestStore<
     if (!request) {
       return;
     }
-    this.resetRequestState(request);
-    this.disposeHiddenAttempt(request);
+    resetVideoScreenshotPreparationRequestState(request);
+    disposeVideoScreenshotHiddenAttempt(request);
     this.compact(captureId, request);
     this.emit();
   }
@@ -81,8 +68,8 @@ export class VideoScreenshotPreparationRequestStore<
       if (!request.tracked || pendingIds.has(captureId)) {
         continue;
       }
-      this.resetRequestState(request);
-      this.disposeHiddenAttempt(request);
+      resetVideoScreenshotPreparationRequestState(request);
+      disposeVideoScreenshotHiddenAttempt(request);
       this.compact(captureId, request);
       didChange = true;
     }
@@ -178,21 +165,11 @@ export class VideoScreenshotPreparationRequestStore<
     captureId: string,
     options: VideoScreenshotHiddenAttemptFailureOptions
   ): VideoScreenshotHiddenAttemptFailureResult {
-    const request = this.requests.get(captureId);
-    if (!request || !request.tracked) {
-      return 'ignored';
-    }
-
-    if (request.hiddenAttemptCount >= Math.max(1, options.maxAttempts)) {
-      request.hiddenTerminalFailed = true;
-      request.hiddenRetryAvailableAt = 0;
+    const result = recordVideoScreenshotHiddenAttemptFailure(this.requests.get(captureId), options);
+    if (result !== 'ignored') {
       this.emit();
-      return 'terminal';
     }
-
-    request.hiddenRetryAvailableAt = Math.max(0, options.retryAvailableAt);
-    this.emit();
-    return 'retry';
+    return result;
   }
 
   abortHiddenAttempt(captureId: string): void {
@@ -200,7 +177,7 @@ export class VideoScreenshotPreparationRequestStore<
     if (!request || !request.hiddenAttempt) {
       return;
     }
-    this.disposeHiddenAttempt(request);
+    disposeVideoScreenshotHiddenAttempt(request);
     this.compact(captureId, request);
     this.emit();
   }
@@ -214,7 +191,7 @@ export class VideoScreenshotPreparationRequestStore<
       if (request.hiddenAttempt?.sourceVideo !== sourceVideo) {
         continue;
       }
-      this.disposeHiddenAttempt(request);
+      disposeVideoScreenshotHiddenAttempt(request);
       this.compact(captureId, request);
       didChange = true;
     }
@@ -236,8 +213,8 @@ export class VideoScreenshotPreparationRequestStore<
         request.visibleAttempted ||
         request.visibleInFlight ||
         Boolean(request.hiddenAttempt);
-      this.resetRequestState(request);
-      this.disposeHiddenAttempt(request);
+      resetVideoScreenshotPreparationRequestState(request);
+      disposeVideoScreenshotHiddenAttempt(request);
       this.requests.delete(captureId);
     }
     if (didChange) {
@@ -268,17 +245,7 @@ export class VideoScreenshotPreparationRequestStore<
   }
 
   canStartHiddenAttempt(captureId: string, now: number): boolean {
-    const request = this.requests.get(captureId);
-    if (!request) {
-      return false;
-    }
-    return (
-      request.tracked &&
-      request.hiddenAttempt === null &&
-      !request.hiddenTerminalFailed &&
-      !request.visibleInFlight &&
-      request.hiddenRetryAvailableAt <= now
-    );
+    return canStartVideoScreenshotHiddenAttempt(this.requests.get(captureId), now);
   }
 
   getHiddenAttemptCount(): number {
@@ -292,22 +259,7 @@ export class VideoScreenshotPreparationRequestStore<
   }
 
   getNextHiddenRetryAt(now: number): number | null {
-    let retryAt: number | null = null;
-    for (const request of this.requests.values()) {
-      if (
-        !request.tracked ||
-        request.hiddenAttempt !== null ||
-        request.hiddenTerminalFailed ||
-        request.hiddenRetryAvailableAt <= now
-      ) {
-        continue;
-      }
-      retryAt =
-        retryAt === null
-          ? request.hiddenRetryAvailableAt
-          : Math.min(retryAt, request.hiddenRetryAvailableAt);
-    }
-    return retryAt;
+    return getNextVideoScreenshotHiddenRetryAt(this.requests.values(), now);
   }
 
   private ensureRequest(captureId: string): VideoScreenshotPreparationRequestState<TAttempt> {
@@ -315,37 +267,9 @@ export class VideoScreenshotPreparationRequestStore<
     if (existing) {
       return existing;
     }
-    const created: VideoScreenshotPreparationRequestState<TAttempt> = {
-      tracked: false,
-      hiddenAttemptCount: 0,
-      hiddenRetryAvailableAt: 0,
-      hiddenTerminalFailed: false,
-      explicitVisible: false,
-      visibleAttempted: false,
-      visibleInFlight: false,
-      hiddenAttempt: null
-    };
+    const created = createVideoScreenshotPreparationRequestState<TAttempt>();
     this.requests.set(captureId, created);
     return created;
-  }
-
-  private disposeHiddenAttempt(request: VideoScreenshotPreparationRequestState<TAttempt>): void {
-    request.hiddenAttempt?.scope.dispose();
-    request.hiddenAttempt = null;
-  }
-
-  private resetRequestState(request: VideoScreenshotPreparationRequestState<TAttempt>): void {
-    request.tracked = false;
-    this.resetHiddenProgress(request);
-    request.explicitVisible = false;
-    request.visibleAttempted = false;
-    request.visibleInFlight = false;
-  }
-
-  private resetHiddenProgress(request: VideoScreenshotPreparationRequestState<TAttempt>): void {
-    request.hiddenAttemptCount = 0;
-    request.hiddenRetryAvailableAt = 0;
-    request.hiddenTerminalFailed = false;
   }
 
   private compact(
@@ -357,16 +281,7 @@ export class VideoScreenshotPreparationRequestStore<
     if (!request) {
       return;
     }
-    if (
-      request.tracked ||
-      request.hiddenAttemptCount > 0 ||
-      request.hiddenRetryAvailableAt > 0 ||
-      request.hiddenTerminalFailed ||
-      request.explicitVisible ||
-      request.visibleAttempted ||
-      request.visibleInFlight ||
-      request.hiddenAttempt
-    ) {
+    if (shouldRetainVideoScreenshotPreparationRequest(request)) {
       return;
     }
     this.requests.delete(captureId);
